[BE - Edit] Change minimum_item_origin type to BigInteger

+++ budgetin/api/models/coa_model.py
@@ -8,4 +8,4 @@ class Coa(SoftDeleteModel, TimestampModel, UserTrackModel):
     definition = models.CharField(max_length=1000, blank=True, null=True)
     hyperion_name = models.CharField(max_length=200)
     is_capex = models.BooleanField(default=False)
-    minimum_item_origin = models.IntegerField(blank=True, null=True)
+    minimum_item_origin = models.BigIntegerField(blank=True, null=True)
